fix: Fix blank line throwing exception
RCA: If the file the program reads has blank lines, the regex finds
nothing to be the change_id and when we try to access the list result at
index 0, it throws a list out of bounds exception.

SOL: Don't call the apply_individual method if the url received is
empty.

Signed-off-by: Malbolge <victor.gomes@icomp.ufam.edu.br>