fix: Check if branch exists before creating it
RCA: If the users chooses to create a new branch where a cherry-pick
will be applied and there are multiple cherry-picks to the same repo,
the program would try to create the same branch that is already created
by the first time it visited this path.

SOL: Check if the branch is already created before creating it. Also,
try to avoid race conditions by the threads accessing the same data
structure.

Signed-off-by: Malbolge <victor.gomes@icomp.ufam.edu.br>