fix: Check if branch existis before creating it
RCA: The current check of branch only works if the branch has previously
been created by this script in a current application. If the branch
existed prior to the execution of the scripts, it wound't know.

SOL: Change the way to check if a branch exists by running a subsprocess
with a git command.

Signed-off-by: Malbolge <victor.gomes@icomp.ufam.edu.br>