fix: Fix checking branch not working
RCA: The subprocess does not need the 'shell=True' parameter when we are
sending a list of commands. This causes the subprocess spawn a shell
process and then execute the command in this process. The return, in
this case, will always be the returned of the execution of the shell
process and not from the command passed, which is not what we want.

SOL: Remove the 'shell=True' parameter from the _run_command_list
method.

Signed-off-by: Malbolge <victor.gomes@icomp.ufam.edu.br>